+++ src/components/Map.vue
@@ -498,6 +498,9 @@ export default {
               showlegend: false,
               name: "NWIS Gage Data",
               hovertemplate: "%{x}<br>Gage height: %{y} feet<extra></extra>",
+              font: {
+                family: "Public Sans, sans-serif",
+              },
             },
           ];
 
@@ -513,6 +516,9 @@ export default {
             autosize: false,
             width: 400,
             height: 400,
+            font: {
+              family: "Public Sans, sans-serif",
+            },
             yaxis: {
               title: "Gage Height, feet",
               titlefont: { size: 12 },
@@ -530,7 +536,6 @@ export default {
               font: {
                 size: 12,
                 color: "rgba(0,0,0,0.6)",
-                family: "Public Sans, sans-serif",
               },
               x: 0.05,
             },
@@ -542,6 +547,11 @@ export default {
             },
             legend: false,
             annotations: plotlyAnnotations,
+            hoverlabel: {
+              font: {
+                family: "Public Sans, sans-serif",
+              },
+            },
           };
 
           // Make chart responsive and modebar always visible
@@ -666,6 +676,9 @@ export default {
             showlegend: true,
             name: "<b>NWIS Gage Data</b>",
             hovertemplate: "%{x}<br>Gage height: %{y} feet<extra></extra>",
+            font: {
+              family: "Public Sans, sans-serif",
+            },
           },
         ];
 
@@ -720,6 +733,9 @@ export default {
             name: "<b>Threshold</b>",
             // Tooltip
             hovertemplate: "%{fullData.name}: %{y} feet<extra></extra>",
+            font: {
+              family: "Public Sans, sans-serif",
+            },
           });
 
           // Create labels
@@ -732,6 +748,7 @@ export default {
             showarrow: true,
             arrowhead: 0,
             font: {
+              family: "Public Sans, sans-serif",
               size: 11,
             },
             ax: ax,
@@ -782,6 +799,9 @@ export default {
           let layout = {
             autosize: false,
             width: 600,
+            font: {
+              family: "Public Sans, sans-serif",
+            },
             yaxis: {
               title: "Gage Height, feet",
               titlefont: { size: 12 },
@@ -799,7 +819,6 @@ export default {
               font: {
                 size: 12,
                 color: "rgba(0,0,0,0.6)",
-                family: "Public Sans, sans-serif",
               },
               x: 0.05,
             },
@@ -807,7 +826,6 @@ export default {
               x: 0.25,
               y: -0.4,
               font: {
-                family: "sans-serif",
                 size: 12,
               },
               orientation: "h",
@@ -820,6 +838,11 @@ export default {
               pad: 4,
             },
             annotations: plotlyAnnotations,
+            hoverlabel: {
+              font: {
+                family: "Public Sans, sans-serif",
+              },
+            },
           };
 
           // Make chart responsive and modebar always visible