+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Changed icons from boxes to circles
+- Changed graphs font to Public Sans
 - Made legend collapsible
 
 ### Fixed